Add scaling factor for data distribution particle counts
Add scaling paramter to tests to scale particle count in order
to create memory footprints that more accurately reflect a data
distribution.

Data distribution inputs are in particle counts.
Particles are 32-byte structures so a data distribution measured in
bytes needs to be scaled down so the particals instiated match
the actual data footprint, in multiples of 32-byte particles.