Fix the data dist benchmark binary name to use "var" naming convention
Change the deployed binary to "write_var_data_dist" so it matches
the configured test reference in h5bench.
The binary name needs to match the test name so that it can be called
by the hbench wrapper.
This follows convention of "write_var_normal_dist".
Update code to log the correct benchmark name.